% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_phs}
\alias{parse_phs}
\title{Get phs identifiers from a dbGaP-provided Study Request List}
\usage{
parse_phs(html_path)
}
\arguments{
\item{html_path}{filepath to the html page downloaded from dbGaP}
}
\value{
a list whose elements are each a list of phs identifiers with one
list per projects
}
\description{
Get phs identifiers from a dbGaP-provided Study Request List
}
\examples{
\dontrun{
html_path <-
system.file("extdata", "example.html", package = "phsparsr", mustWork = TRUE)
ids <- parse_phs(html_path)
}
}
